# Use r-tree for LineString's IsSimple