This fixes the "worst case scenario" for this algorithm, which occurs
when the line segments in the LineString are stacked on top of each
other vertically (vertical zig zag).

Performance change:

```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkLineStringIsSimpleZigZag/10-4        1499          3683          +145.70%
BenchmarkLineStringIsSimpleZigZag/100-4       72693         93119         +28.10%
BenchmarkLineStringIsSimpleZigZag/1000-4      5991268       1265259       -78.88%
BenchmarkLineStringIsSimpleZigZag/10000-4     585784268     16355471      -97.21%
```

Worse for smaller inputs, but quickly catches up for inputs > 100. Significantly better complexity class (quadratic vs linearithmic). I suspect there are some "large" constants for the new solution related to memory management in the rtree.